New function in MogileFS::Admin: get_stats.
   my $stats = $mgd->get_stats();

Can take as parameters a list of the statistics to gather.  Right now only
the 'replication' statistic is available.  With no parameters, or with the
'all' parameter, it will gather all available statistics.

The replication statistic provides a hashref of information that tells you
the number of files with each devcount, broken down by domain and class.
At a glance you can tell how many times each file has been replicated.
This information can be combined with the information returned by the
get_domains method to determine how far behind the replicator is.


git-svn-id: http://code.sixapart.com/svn/mogilefs/trunk@66 f67b2e87-0811-0410-a7e0-dd94e48410d6