Add new MogileFS::Admin command: change_device_state( host, deviceid, state ).
This enables you to set the state of a device.  Valid states are alive,
down, and dead.  This also enforces that you can't go from dead to
alive--you have to stop at down first.  Also, this verifies that the
selected device does indeed belong to the host you've specified, so there's
less of a chance for error.


git-svn-id: http://code.sixapart.com/svn/mogilefs/trunk@68 f67b2e87-0811-0410-a7e0-dd94e48410d6